fixed sdl initialization and unnecessary copying

+++ include/Image.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "Pixel.h"
+#include <cstdint>
 #include <string>
 #include <vector>
 
@@ -7,6 +8,6 @@ struct Image {
   int width = 0;
   int height = 0;
   std::string ppmType;
-  uint8_t maxVal = 255;
+  uint16_t maxVal = 255;
   std::vector<Pixel> pixelData;
 };
