+++ src/Parser.cpp
@@ -75,7 +75,7 @@ std::optional<Image> Parser::ParseFile(const std::string &filePath,
     ss.str(line);
   }
   ss >> maxValInt;
-  image.maxVal = static_cast<uint8_t>(maxValInt);
+  image.maxVal = static_cast<uint16_t>(maxValInt);
   if (verbose) {
     std::cout << "File pre-parsing has been completed.\nPPM image type: "
               << image.ppmType << "\nImage size: " << image.width << ","
